requirements.txt
streamlit
openai
pandas
numpy
python-dotenv
thefuzz
kaggle
tabulate
